+++ mobile/android/android-components/components/feature/downloads/src/test/java/mozilla/components/feature/downloads/manager/FetchDownloadManagerTest.kt
@@ -14,9 +14,9 @@ import android.content.Intent
 import androidx.localbroadcastmanager.content.LocalBroadcastManager
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import mozilla.components.browser.state.state.content.DownloadState
+import mozilla.components.browser.state.store.BrowserStore
 import mozilla.components.concept.fetch.Client
 import mozilla.components.feature.downloads.AbstractFetchDownloadService
-import mozilla.components.feature.downloads.AbstractFetchDownloadService.Companion.EXTRA_DOWNLOAD
 import mozilla.components.feature.downloads.AbstractFetchDownloadService.Companion.EXTRA_DOWNLOAD_STATUS
 import mozilla.components.support.test.any
 import mozilla.components.support.test.mock
@@ -27,12 +27,12 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import org.mockito.Mockito.verify
 import mozilla.components.feature.downloads.AbstractFetchDownloadService.DownloadJobStatus
-import org.junit.Assert.assertFalse
+import mozilla.components.support.test.libstate.ext.waitUntilIdle
 import org.junit.Assert.assertTrue
 import org.junit.Assert.assertNull
 import org.junit.Assert.assertNotNull
 import org.junit.Assert.assertEquals
-import org.mockito.Mockito.times
+import org.mockito.Mockito.never
 
 @RunWith(AndroidJUnit4::class)
 class FetchDownloadManagerTest {
@@ -41,17 +41,19 @@ class FetchDownloadManagerTest {
     private lateinit var service: MockDownloadService
     private lateinit var download: DownloadState
     private lateinit var downloadManager: FetchDownloadManager<MockDownloadService>
+    private lateinit var store: BrowserStore
 
     @Before
     fun setup() {
         broadcastManager = LocalBroadcastManager.getInstance(testContext)
         service = MockDownloadService()
+        store = BrowserStore()
         download = DownloadState(
             "http://ipv4.download.thinkbroadband.com/5MB.zip",
             "", "application/zip", 5242880,
             "Mozilla/5.0 (Linux; Android 7.1.1) AppleWebKit/537.36 (KHTML, like Gecko) Version/4.0 Focus/8.0 Chrome/69.0.3497.100 Mobile Safari/537.36"
         )
-        downloadManager = FetchDownloadManager(testContext, MockDownloadService::class, broadcastManager)
+        downloadManager = FetchDownloadManager(testContext, store, MockDownloadService::class, broadcastManager)
     }
 
     @Test(expected = SecurityException::class)
@@ -60,80 +62,64 @@ class FetchDownloadManagerTest {
     }
 
     @Test
-    fun `calling download must download the file`() {
+    fun `calling download must queue the download`() {
         val context: Context = mock()
-        downloadManager = FetchDownloadManager(context, MockDownloadService::class, broadcastManager)
-        var downloadCompleted = false
+        downloadManager = FetchDownloadManager(context, store, MockDownloadService::class, broadcastManager)
+        var downloadStopped = false
 
-        downloadManager.onDownloadStopped = { _, _, _ -> downloadCompleted = true }
+        downloadManager.onDownloadStopped = { _, _, _ -> downloadStopped = true }
 
         grantPermissions()
 
+        assertTrue(store.state.queuedDownloads.isEmpty())
         val id = downloadManager.download(download)!!
-
-        verify(context).startService(any())
+        store.waitUntilIdle()
+        assertEquals(download, store.state.queuedDownloads[download.id])
 
         notifyDownloadCompleted(id)
-
-        assertTrue(downloadCompleted)
+        assertTrue(downloadStopped)
     }
 
     @Test
     fun `calling tryAgain starts the download again`() {
         val context: Context = mock()
-        downloadManager = FetchDownloadManager(context, MockDownloadService::class, broadcastManager)
-        var downloadCompleted = false
+        downloadManager = FetchDownloadManager(context, store, MockDownloadService::class, broadcastManager)
+        var downloadStopped = false
 
-        downloadManager.onDownloadStopped = { _, _, _ -> downloadCompleted = true }
+        downloadManager.onDownloadStopped = { _, _, _ -> downloadStopped = true }
 
         grantPermissions()
 
         val id = downloadManager.download(download)!!
+        store.waitUntilIdle()
+        notifyDownloadFailed(id)
+        assertTrue(downloadStopped)
 
+        downloadStopped = false
+        downloadManager.tryAgain(id)
         verify(context).startService(any())
         notifyDownloadCompleted(id)
-        assertTrue(downloadCompleted)
-
-        downloadCompleted = false
-
-        downloadManager.tryAgain(id)
-
-        verify(context, times(2)).startService(any())
-        notifyDownloadCompleted(id)
-        assertTrue(downloadCompleted)
+        assertTrue(downloadStopped)
     }
 
     @Test
     fun `try again should not crash when download does not exist`() {
         val context: Context = mock()
-        downloadManager = FetchDownloadManager(context, MockDownloadService::class, broadcastManager)
-        var downloadCompleted = false
-
-        downloadManager.onDownloadStopped = { _, _, _ -> downloadCompleted = true }
+        downloadManager = FetchDownloadManager(context, store, MockDownloadService::class, broadcastManager)
 
         grantPermissions()
-
         val id = downloadManager.download(download)!!
 
-        verify(context).startService(any())
-        notifyDownloadCompleted(id)
-        assertTrue(downloadCompleted)
-
-        downloadCompleted = false
         downloadManager.tryAgain(id + 1)
-        assertFalse(downloadCompleted)
-        verify(context, times(1)).startService(any())
+        verify(context, never()).startService(any())
     }
 
     @Test
     fun `trying to download a file with invalid protocol must NOT triggered a download`() {
-
         val invalidDownload = download.copy(url = "ftp://ipv4.download.thinkbroadband.com/5MB.zip")
-
         grantPermissions()
 
         val id = downloadManager.download(invalidDownload)
-
         assertNull(id)
     }
 
@@ -148,7 +134,7 @@ class FetchDownloadManagerTest {
 
     @Test
     fun `sendBroadcast with valid downloadID must call onDownloadStopped after download`() {
-        var downloadCompleted = false
+        var downloadStopped = false
         var downloadStatus: DownloadJobStatus? = null
         val downloadWithFileName = download.copy(fileName = "5MB.zip")
 
@@ -156,24 +142,46 @@ class FetchDownloadManagerTest {
 
         downloadManager.onDownloadStopped = { _, _, status ->
             downloadStatus = status
-            downloadCompleted = true
+            downloadStopped = true
         }
 
         val id = downloadManager.download(
             downloadWithFileName,
             cookie = "yummy_cookie=choco"
         )!!
+        store.waitUntilIdle()
 
         notifyDownloadCompleted(id)
-
-        assertTrue(downloadCompleted)
-
+        assertTrue(downloadStopped)
         assertEquals(DownloadJobStatus.COMPLETED, downloadStatus)
     }
 
     @Test
+    fun `sendBroadcast with completed download removes queued download from store`() {
+        var downloadStatus: DownloadJobStatus? = null
+        val downloadWithFileName = download.copy(fileName = "5MB.zip")
+        grantPermissions()
+
+        downloadManager.onDownloadStopped = { _, _, status ->
+            downloadStatus = status
+        }
+
+        val id = downloadManager.download(
+            downloadWithFileName,
+            cookie = "yummy_cookie=choco"
+        )!!
+        store.waitUntilIdle()
+        assertEquals(downloadWithFileName, store.state.queuedDownloads[downloadWithFileName.id])
+
+        notifyDownloadCompleted(id)
+        store.waitUntilIdle()
+        assertEquals(DownloadJobStatus.COMPLETED, downloadStatus)
+        assertTrue(store.state.queuedDownloads.isEmpty())
+    }
+
+    @Test
     fun `onReceive properly gets download object form sendBroadcast`() {
-        var downloadCompleted = false
+        var downloadStopped = false
         var downloadStatus: DownloadJobStatus? = null
         var downloadName = ""
         var downloadSize = 0L
@@ -183,26 +191,32 @@ class FetchDownloadManagerTest {
 
         downloadManager.onDownloadStopped = { download, _, status ->
             downloadStatus = status
-            downloadCompleted = true
+            downloadStopped = true
             downloadName = download.fileName ?: ""
             downloadSize = download.contentLength ?: 0
         }
 
         val id = downloadManager.download(downloadWithFileName)!!
+        store.waitUntilIdle()
+        notifyDownloadCompleted(id)
 
-        notifyDownloadCompleted(id, downloadWithFileName)
-
-        assertTrue(downloadCompleted)
+        assertTrue(downloadStopped)
         assertEquals("5MB.zip", downloadName)
         assertEquals(5L, downloadSize)
         assertEquals(DownloadJobStatus.COMPLETED, downloadStatus)
     }
 
-    private fun notifyDownloadCompleted(id: Long, download: DownloadState = DownloadState(url = "")) {
+    private fun notifyDownloadFailed(id: Long) {
+        val intent = Intent(ACTION_DOWNLOAD_COMPLETE)
+        intent.putExtra(EXTRA_DOWNLOAD_ID, id)
+        intent.putExtra(EXTRA_DOWNLOAD_STATUS, DownloadJobStatus.FAILED)
+        broadcastManager.sendBroadcast(intent)
+    }
+
+    private fun notifyDownloadCompleted(id: Long) {
         val intent = Intent(ACTION_DOWNLOAD_COMPLETE)
         intent.putExtra(EXTRA_DOWNLOAD_ID, id)
         intent.putExtra(EXTRA_DOWNLOAD_STATUS, DownloadJobStatus.COMPLETED)
-        intent.putExtra(EXTRA_DOWNLOAD, download)
 
         broadcastManager.sendBroadcast(intent)
     }
@@ -213,5 +227,6 @@ class FetchDownloadManagerTest {
 
     class MockDownloadService : AbstractFetchDownloadService() {
         override val httpClient: Client = mock()
+        override val store: BrowserStore = mock()
     }
 }