+++ mobile/android/android-components/samples/browser/src/main/java/org/mozilla/samples/browser/downloads/DownloadService.kt
@@ -4,9 +4,11 @@
 
 package org.mozilla.samples.browser.downloads
 
+import mozilla.components.browser.state.store.BrowserStore
 import mozilla.components.feature.downloads.AbstractFetchDownloadService
 import org.mozilla.samples.browser.ext.components
 
 class DownloadService : AbstractFetchDownloadService() {
     override val httpClient by lazy { components.client }
+    override val store: BrowserStore by lazy { components.store }
 }